Profiles: updating `identifier` and `required` docs
I hear our documentation for identifiers is confusing. I've made some
minor changes to (hopefully) make their use in a route less confusing.
Those changes are an attempt to make the `-` more visible to API
consumers.

The changes to the required fields are a clarification on when they're
actually required. They're only required for POSTs, and originally there
was only one line in the PUT documentation saying that required fields
(despite their name) aren't actually required. Mix that with the
boldface next to those fields in the profile object description and you
have a recipe for some trib troubs.